requirements/dev.txt
# Only development-specific requirements but imports the common ones too.
-r common.txt

black == 23.3.0
django-debug-toolbar == 4.0.0
django-extensions == 3.2.1
model-bakery == 1.11.0
python-dotenv == 1.0.0
